Data.As

Description

As is morally identical to Tagged from the tagged package, except the latter is rendered unusable for serialization tagging by existing libraries.

Documentation

data As (s :: Symbol) a

Tagged type.

unAs :: forall (s :: Symbol) a. As s a -> a

Remove the tag.