Data.Country

Description

Country represents an encoding-agnostic ISO 3166-1 country, identified as a 3-tuple (alpha-2, alpha-3, numeric) of its standard representations.

Note: numeric country codes are not numbers and should not be treated as such. Serialization to numbers is thus only provided for compatibility purposes.

Laziness

Evaluating a Country to WHNF evaluates it to normal form.

Itself

data Country

ISO 3166-1 country.

Type class shim

newtype As (s :: Symbol) a

Tagged type.

Constructors

As a

unAs :: forall (s :: Symbol) a. As s a -> a

Remove the tag.

Construct

mkCountry

Arguments

:: ByteString
Alpha-2 code

-> ByteString
Alpha-3 code

-> ByteString
Numeric code

-> Maybe Country

Construct a Country out of its three ISO 3166-1 representations.

Encode

Alpha-2

encodeAlpha2 :: Country -> Text

Encode a country as an alpha-2 code into an ASCII string.

Alpha-3

encodeAlpha3 :: Country -> Text

Encode a country as an alpha-3 code into an ASCII string.

Numeric

encodeNumeric :: Country -> Text

Encode a country as a numeric code into an ASCII string.

Integer representation

encodeNumericAsInt :: Country -> Int

Encode a country as a numeric code into an integer.

Decode

Alpha-2

decodeAlpha2 :: ByteString -> Either Text Country

decodeAlpha2' over the latest revision.

decodeAlpha2' :: Revision -> ByteString -> Either Text Country

Decode an alpha-2 country code from an ASCII string.

Alpha-3

decodeAlpha3 :: ByteString -> Either Text Country

decodeAlpha3' over the latest revision.

decodeAlpha3' :: Revision -> ByteString -> Either Text Country

Decode an alpha-3 country code from an ASCII string.

Numeric

decodeNumeric :: ByteString -> Either Text Country

decodeNumeric' over the latest revision.

decodeNumeric' :: Revision -> ByteString -> Either Text Country

Decode a numeric country code from an ASCII string.

Integer representation

decodeNumericFromInt :: Int -> Either Text Country

decodeNumericFromInt' over the latest revision.

decodeNumericFromInt' :: Revision -> Int -> Either Text Country

Decode a numeric country code from an integer.