Data.Country.Aeson

Description

Country code serialization functions for aeson.

Encode

Alpha-2

toAlpha2 :: Country -> Value

Encode a country as an alpha-2 code into a JSON string.

Alpha-3

toAlpha3 :: Country -> Value

Encode a country as an alpha-3 code into a JSON string.

Numeric

toNumeric :: Country -> Value

Encode a country as a numeric code into a JSON string.

Integer representation

toNumericAsNumber :: Country -> Value

Encode a country as a numeric code into a JSON number.

Decode

Alpha-2

parseAlpha2 :: Value -> Parser Country

parseAlpha2' over the latest revision.

parseAlpha2' :: Revision -> Value -> Parser Country

Decode an alpha-2 country code from a JSON string.

Alpha-3

parseAlpha3 :: Value -> Parser Country

parseAlpha3' over the latest revision.

parseAlpha3' :: Revision -> Value -> Parser Country

Decode an alpha-3 country code from a JSON string.

Numeric

parseNumeric :: Value -> Parser Country

parseNumeric' over the latest revision.

parseNumeric' :: Revision -> Value -> Parser Country

Decode a numeric country code from a JSON string.

Integer representation

parseNumericFromNumber :: Value -> Parser Country

parseNumericFromNumber' over the latest revision.

parseNumericFromNumber' :: Revision -> Value -> Parser Country

Decode a numeric country code from a JSON number.