Data.Country.PostgreSQL

Description

Country code serialization functions for postgresql-simple.

Encode

Alpha-2

toAlpha2 :: Country -> Action

Encode a country as an alpha-2 code into an SQL query string.

Alpha-3

toAlpha3 :: Country -> Action

Encode a country as an alpha-3 code into an SQL query string.

Numeric

toNumeric :: Country -> Action

Encode a country as an numeric code into an SQL query string representing a text value.

Integer representation

toNumericAsInt :: Country -> Action

Encode a country as an numeric code into an SQL query string representing an integer.

Decode

Alpha-2

fromAlpha2 :: FieldParser Country

fromAlpha2' over the latest revision.

fromAlpha2' :: Revision -> FieldParser Country

Decode an alpha-2 country from an SQL value.

Alpha-3

fromAlpha3 :: FieldParser Country

fromAlpha3' over the latest revision.

fromAlpha3' :: Revision -> FieldParser Country

Decode an alpha-3 country from an SQL value.

Numeric

fromNumeric :: FieldParser Country

fromNumeric' over the latest revision.

fromNumeric' :: Revision -> FieldParser Country

Decode a numeric country from a textual SQL value.

Integer representation

fromNumericAsInt :: FieldParser Country

fromNumericAsInt' over the latest revision.

fromNumericAsInt' :: Revision -> FieldParser Country

Decode a numeric country from a numeric SQL value.