Data.Currency.Aeson

Description

Currency code serialization functions for aeson.

Encode

Alpha

toAlpha :: Currency -> Value

Encode a currency as an alpha code into a JSON string.

Numeric

toNumeric :: Currency -> Value

Encode a currency as a numeric code into a JSON string.

Integer representation

toNumericAsNumber :: Currency -> Value

Encode a currency as a numeric code into a JSON number.

Decode

Alpha

parseAlpha :: Value -> Parser Currency

parseAlpha' over the latest revision.

parseAlpha' :: Revision -> Value -> Parser Currency

Decode an alpha currency code from a JSON string.

Numeric

parseNumeric :: Value -> Parser Currency

parseNumeric' over the latest revision.

parseNumeric' :: Revision -> Value -> Parser Currency

Decode a numeric currency code from a JSON string.

Integer representation

parseNumericFromNumber :: Value -> Parser Currency

parseNumericFromNumber' over the latest revision.

parseNumericFromNumber' :: Revision -> Value -> Parser Currency

Decode a numeric currency code from a JSON number.