Data.Currency.PostgreSQL

Description

Currency code serialization functions for postgresql-simple.

Encode

Alpha

toAlpha :: Currency -> Action

Encode a currency as an alpha code into an SQL query string.

Numeric

toNumeric :: Currency -> Action

Encode a currency as an numeric code into an SQL query string representing a text value.

Integer representation

toNumericAsInt :: Currency -> Action

Encode a currency as an numeric code into an SQL query string representing an integer.

Decode

Alpha

fromAlpha :: FieldParser Currency

fromAlpha' over the latest revision.

fromAlpha' :: Revision -> FieldParser Currency

Decode an alpha currency from an SQL value.

Numeric

fromNumeric :: FieldParser Currency

fromNumeric' over the latest revision.

fromNumeric' :: Revision -> FieldParser Currency

Decode a numeric currency from a textual SQL value.

Integer representation

fromNumericAsInt :: FieldParser Currency

fromNumericAsInt' over the latest revision.

fromNumericAsInt' :: Revision -> FieldParser Currency

Decode a numeric currency from a numeric SQL value.