Data.Currency.Revision.Unsafe

Description

Data structure internals, helper operations and unsafe functions.

Itself

newtype Revision

Set of currencies.

Constructors

Revision (IntMap Currency)

Key

newtype Key

Revision key type.

Constructors

Key Word

alphaToKey :: Alpha -> Key

Convert an alpha code to a dictionary key.

numericToKey :: Numeric -> Key

Convert a numeric code to a dictionary key.

Lookup

lookup :: Key -> Revision -> Maybe Currency

Look up a currency at a key in a revision.

Insert

unsafeInsert :: Currency -> Revision -> Revision

Add the representations of a given currency to a revision.

The revision MUST NOT have in it other currencies with the same representation(s).

Delete

unsafeDelete :: Currency -> Revision -> Revision

Remove the representations of a given currency from a revision.

If the revision does not contain in it the given currency, it MUST NOT have in it other currencies with the same representation(s).