Data.Language.Revision

Description

Revision represents a set of languages as defined by the ISO 639 standard at a particular point in time.

All codes within a particular revision are guaranteed to be unique by the standard.

Itself

data Revision

Set of languages.

Construct

empty :: Revision

Revision with no languages.

Query

member :: Language -> Revision -> Bool

Check if the given language is in a revision.

overlaps :: Language -> Revision -> Bool

Check if any of the representations of the given language is in a revision.

Lookup

lookupAlpha2 :: ByteString -> Revision -> Maybe Language

Parse an alpha-2 code from a two-byte ASCII sequence.

lookupAlpha3 :: ByteString -> Revision -> Maybe Language

Parse an alpha-3 code from a three-byte ASCII sequence.

Insert

insert :: Language -> Revision -> Maybe Revision

Add a language to a revision.

Returns Nothing if any other languages in the revision already hold the same representation.

Delete

delete :: Language -> Revision -> Revision

Remove a language from a revision.

List

toList :: Revision -> [Language]

List all languages in a revision.