Data.Language.Revision.Unsafe

Description

Data structure internals, helper operations and unsafe functions.

Itself

newtype Revision

Set of languages.

Constructors

Revision (IntMap Language)

Key

newtype Key

Revision key type.

Constructors

Key Word

alpha2ToKey :: Alpha2 -> Key

Convert an alpha-2 code to a dictionary key.

alpha3ToKey :: Alpha3 -> Key

Convert an alpha-3 code to a dictionary key.

Lookup

lookup :: Key -> Revision -> Maybe Language

Look up a language at a key in a revision.

Insert

unsafeInsert :: Language -> Revision -> Revision

Add the representations of a given language to a revision.

The revision MUST NOT have in it other languages with the same representation(s).

Delete

unsafeDelete :: Language -> Revision -> Revision

Remove the representations of a given language from a revision.

If the revision does not contain in it the given language, it MUST NOT have in it other languages with the same representation(s).